more refactoring

+++ AOC/src/main/java/nl/roykovic/aoc/_2022/day7_filesystem/Directory.java
@@ -46,10 +46,10 @@ public class Directory implements FileSystemItem{
         int size = 0;
 
         for(Directory child: children){
-            size += child.getSize();
+            size += (int) child.getSize();
         }
         for(File file: files){
-            size += file.getSize();
+            size += (int) file.getSize();
         }
 
         return size;
